Put test utility functions into separate file
Also updated JSHint config file and removed all the /*globals ...*/ comments from tests.
