+++ test/builders/exists.test.js
@@ -1,10 +1,12 @@
 /*jshint -W030*/
-/*globals create, testPath, createPaths, deletePaths*/
 'use strict';
 
+var testUtils = require('../support/utils');
 var utils = require('../../lib/utils');
 var Event = Rorschach.Event;
 var Exception = Rorschach.Exception;
+var create = testUtils.create;
+var testPath = testUtils.testPath;
 
 
 describe('ExistsBuilder', function existsTestSuite() {
